# jit: implement function calls in fully assembly