… wasm/internal/spectests/spec_test.go
… wasm/jit/RATIONALE.md
… wasm/jit/engine_test.go
… wasm/jit/engine_tmp_test.go
… wasm/jit/jit_amd64.s
… wasm/jit/jit_amd64_test.go
… wasm/jit/jit_other.go
… wasm/jit/jit_value_location_amd64_test.go
… wasm/jit/testdata/memory.wasm
… wasm/jit/testdata/memory.wat
… wasm/jit/testdata/recursive.wasm
… wasm/jit/testdata/tmp.wasm
… wasm/jit/testdata/tmp.wat