@mathetake 2022-01-25
ok done 80%