@mathetake 2022-01-27
Ok I will mark this ready after polishing docs and observe the benchmark improvements