@mathetake 2022-01-27
Here's the bench result for Recursive fibonacci function written in wat. 
This proves that the cost of function calls was the bottleneck and now we resolved the issue! 

#### before
```
BenchmarkFibonacci
BenchmarkFibonacci/jit
BenchmarkFibonacci/jit-32         	     477	   2480705 ns/op	 1051497 B/op	   21893 allocs/op
BenchmarkFibonacci/interpreter
BenchmarkFibonacci/interpreter-32 	     555	   2242362 ns/op	  350296 B/op	   21893 allocs/op
```

#### after
```
goos: linux
goarch: amd64
pkg: github.com/tetratelabs/wazero/wasm/jit
cpu: AMD Ryzen 9 3950X 16-Core Processor
BenchmarkFibonacci
BenchmarkFibonacci/jit
BenchmarkFibonacci/jit-32         	    7714	    149843 ns/op	      61 B/op	       2 allocs/op
BenchmarkFibonacci/interpreter
BenchmarkFibonacci/interpreter-32 	     564	   2182406 ns/op	  350304 B/op	   21893 allocs/op
```